@@ -1,22 +1,19 @@
-'use strict'
-
-const Libp2p = require('../..')
-const TCP = require('libp2p-tcp')
-const Mplex = require('libp2p-mplex')
-const { NOISE } = require('@chainsafe/libp2p-noise')
-
-const pipe = require('it-pipe')
+import { createLibp2p } from '../../dist/src/index.js'
+import { TCP } from '@libp2p/tcp'
+import { Mplex } from '@libp2p/mplex'
+import { Noise } from '@chainsafe/libp2p-noise'
+import { pipe } from 'it-pipe'
+import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
+import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 
 const createNode = async () => {
-  const node = await Libp2p.create({
+  const node = await createLibp2p({
     addresses: {
       listen: ['/ip4/0.0.0.0/tcp/0']
     },
-    modules: {
-      transport: [TCP],
-      streamMuxer: [Mplex],
-      connEncryption: [NOISE]
-    }
+    transports: [new TCP()],
+    streamMuxers: [new Mplex()],
+    connectionEncryption: [new Noise()]
   })
 
   await node.start()
@@ -30,14 +27,14 @@ const createNode = async () => {
     createNode()
   ])
 
-  await node1.peerStore.addressBook.set(node2.peerId, node2.multiaddrs)
+  await node1.peerStore.addressBook.set(node2.peerId, node2.getMultiaddrs())
 
   node2.handle('/a-protocol', ({ stream }) => {
     pipe(
       stream,
       async function (source) {
         for await (const msg of source) {
-          console.log(msg.toString())
+          console.log(uint8ArrayToString(msg))
         }
       }
     )
@@ -46,7 +43,7 @@ const createNode = async () => {
   const { stream } = await node1.dialProtocol(node2.peerId, '/a-protocol')
 
   await pipe(
-    ['This information is sent out encrypted to the other peer'],
+    [uint8ArrayFromString('This information is sent out encrypted to the other peer')],
     stream
   )
 })();
