@@ -1,14 +1,14 @@
-'use strict'
 
-const path = require('path')
-const { waitForOutput } = require('../utils')
+import path from 'path'
+import { waitForOutput } from '../utils.js'
+import { fileURLToPath } from 'url'
 
-async function test () {
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
+
+export async function test () {
   process.stdout.write('1.js\n')
 
   await waitForOutput('This information is sent out encrypted to the other peer', 'node', [path.join(__dirname, '1.js')], {
     cwd: __dirname
   })
 }
-
-module.exports = test
