@@ -1,43 +1,37 @@
 /* eslint-disable no-console */
-'use strict'
 
-const Libp2p = require('../../')
-const TCP = require('libp2p-tcp')
-const Mplex = require('libp2p-mplex')
-const { NOISE } = require('@chainsafe/libp2p-noise')
-const Bootstrap = require('libp2p-bootstrap')
-
-const bootstrapers = require('./bootstrapers')
+import { createLibp2p } from 'libp2p'
+import { TCP } from '@libp2p/tcp'
+import { Mplex } from '@libp2p/mplex'
+import { Noise } from '@chainsafe/libp2p-noise'
+import { Bootstrap } from '@libp2p/bootstrap'
+import bootstrapers from './bootstrappers.js'
 
 ;(async () => {
-  const node = await Libp2p.create({
+  const node = await createLibp2p({
     addresses: {
       listen: ['/ip4/0.0.0.0/tcp/0']
     },
-    modules: {
-      transport: [TCP],
-      streamMuxer: [Mplex],
-      connEncryption: [NOISE],
-      peerDiscovery: [Bootstrap]
-    },
-    config: {
-      peerDiscovery: {
-        bootstrap: {
-          interval: 60e3,
-          enabled: true,
-          list: bootstrapers
-        }
-      }
-    }
+    transports: [new TCP()],
+    streamMuxers: [new Mplex()],
+    connectionEncryption: [new Noise()],
+    peerDiscovery: [
+      new Bootstrap({
+        interval: 60e3,
+        list: bootstrapers
+      })
+    ]
   })
 
-  node.connectionManager.on('peer:connect', (connection) => {
-    console.log('Connection established to:', connection.remotePeer.toB58String())	// Emitted when a peer has been found
+  node.connectionManager.addEventListener('peer:connect', (evt) => {
+    const connection = evt.detail
+    console.log('Connection established to:', connection.remotePeer.toString())	// Emitted when a peer has been found
   })
 
-  node.on('peer:discovery', (peerId) => {
+  node.addEventListener('peer:discovery', (evt) => {
+    const peer = evt.detail
     // No need to dial, autoDial is on
-    console.log('Discovered:', peerId.toB58String())
+    console.log('Discovered:', peer.id.toString())
   })
 
   await node.start()
