@@ -1,32 +1,23 @@
 /* eslint-disable no-console */
-'use strict'
 
-const Libp2p = require('../../')
-const TCP = require('libp2p-tcp')
-const Mplex = require('libp2p-mplex')
-const { NOISE } = require('@chainsafe/libp2p-noise')
-const { CID } = require('multiformats/cid')
-const KadDHT = require('libp2p-kad-dht')
-
-const all = require('it-all')
-const delay = require('delay')
+import { createLibp2p } from 'libp2p'
+import { TCP } from '@libp2p/tcp'
+import { Mplex } from '@libp2p/mplex'
+import { Noise } from '@chainsafe/libp2p-noise'
+import { CID } from 'multiformats/cid'
+import { KadDHT } from '@libp2p/kad-dht'
+import all from 'it-all'
+import delay from 'delay'
 
 const createNode = async () => {
-  const node = await Libp2p.create({
+  const node = await createLibp2p({
     addresses: {
       listen: ['/ip4/0.0.0.0/tcp/0']
     },
-    modules: {
-      transport: [TCP],
-      streamMuxer: [Mplex],
-      connEncryption: [NOISE],
-      dht: KadDHT
-    },
-    config: {
-      dht: {
-        enabled: true
-      }
-    }
+    transports: [new TCP()],
+    streamMuxers: [new Mplex()],
+    connectionEncryption: [new Noise()],
+    dht: new KadDHT()
   })
 
   await node.start()
@@ -40,8 +31,8 @@ const createNode = async () => {
     createNode()
   ])
 
-  await node1.peerStore.addressBook.set(node2.peerId, node2.multiaddrs)
-  await node2.peerStore.addressBook.set(node3.peerId, node3.multiaddrs)
+  await node1.peerStore.addressBook.set(node2.peerId, node2.getMultiaddrs())
+  await node2.peerStore.addressBook.set(node3.peerId, node3.getMultiaddrs())
 
   await Promise.all([
     node1.dial(node2.peerId),
@@ -54,12 +45,12 @@ const createNode = async () => {
   const cid = CID.parse('QmTp9VkYvnHyrqKQuFPiuZkiX9gPcqj6x5LJ1rmWuSySnL')
   await node1.contentRouting.provide(cid)
 
-  console.log('Node %s is providing %s', node1.peerId.toB58String(), cid.toString())
+  console.log('Node %s is providing %s', node1.peerId.toString(), cid.toString())
 
   // wait for propagation
   await delay(300)
 
   const providers = await all(node3.contentRouting.findProviders(cid, { timeout: 3000 }))
 
-  console.log('Found provider:', providers[0].id.toB58String())
+  console.log('Found provider:', providers[0].id.toString())
 })();
