@@ -1,22 +1,19 @@
-'use strict'
-
-const Libp2p = require('../../')
-const TCP = require('libp2p-tcp')
-const MPLEX = require('libp2p-mplex')
-const { NOISE } = require('@chainsafe/libp2p-noise')
-
-const pipe = require('it-pipe')
+import { createLibp2p } from 'libp2p'
+import { TCP } from '@libp2p/tcp'
+import { Mplex } from '@libp2p/mplex'
+import { Noise } from '@chainsafe/libp2p-noise'
+import { pipe } from 'it-pipe'
+import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
+import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
 
 const createNode = async () => {
-  const node = await Libp2p.create({
+  const node = await createLibp2p({
     addresses: {
       listen: ['/ip4/0.0.0.0/tcp/0']
     },
-    modules: {
-      transport: [TCP],
-      streamMuxer: [MPLEX],
-      connEncryption: [NOISE]
-    }
+    transports: [new TCP()],
+    streamMuxers: [new Mplex()],
+    connectionEncryption: [new Noise()]
   })
 
   await node.start()
@@ -31,7 +28,7 @@ const createNode = async () => {
   ])
 
   // Add node's 2 data to the PeerStore
-  await node1.peerStore.addressBook.set(node2.peerId, node2.multiaddrs)
+  await node1.peerStore.addressBook.set(node2.peerId, node2.getMultiaddrs())
 
   // exact matching
   node2.handle('/your-protocol', ({ stream }) => {
@@ -39,7 +36,7 @@ const createNode = async () => {
       stream,
       async function (source) {
         for await (const msg of source) {
-          console.log(msg.toString())
+          console.log(uint8ArrayToString(msg))
         }
       }
     )
@@ -56,7 +53,7 @@ const createNode = async () => {
       stream,
       async function (source) {
         for await (const msg of source) {
-          console.log(msg.toString())
+          console.log(uint8ArrayToString(msg))
         }
       }
     )
@@ -65,7 +62,7 @@ const createNode = async () => {
 
   const { stream } = await node1.dialProtocol(node2.peerId, ['/your-protocol'])
   await pipe(
-    ['my own protocol, wow!'],
+    [uint8ArrayFromString('my own protocol, wow!')],
     stream
   )
 
