@@ -1,22 +1,19 @@
-'use strict'
-
-const Libp2p = require('../../')
-const TCP = require('libp2p-tcp')
-const MPLEX = require('libp2p-mplex')
-const { NOISE } = require('@chainsafe/libp2p-noise')
-
-const pipe = require('it-pipe')
+import { createLibp2p } from 'libp2p'
+import { TCP } from '@libp2p/tcp'
+import { Mplex } from '@libp2p/mplex'
+import { Noise } from '@chainsafe/libp2p-noise'
+import { pipe } from 'it-pipe'
+import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
+import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
 
 const createNode = async () => {
-  const node = await Libp2p.create({
+  const node = await createLibp2p({
     addresses: {
       listen: ['/ip4/0.0.0.0/tcp/0']
     },
-    modules: {
-      transport: [TCP],
-      streamMuxer: [MPLEX],
-      connEncryption: [NOISE]
-    }
+    transports: [new TCP()],
+    streamMuxers: [new Mplex()],
+    connectionEncryption: [new Noise()]
   })
 
   await node.start()
@@ -31,14 +28,14 @@ const createNode = async () => {
   ])
 
   // Add node's 2 data to the PeerStore
-  await node1.peerStore.addressBook.set(node2.peerId, node2.multiaddrs)
+  await node1.peerStore.addressBook.set(node2.peerId, node2.getMultiaddrs())
 
   node2.handle(['/a', '/b'], ({ protocol, stream }) => {
     pipe(
       stream,
       async function (source) {
         for await (const msg of source) {
-          console.log(`from: ${protocol}, msg: ${msg.toString()}`)
+          console.log(`from: ${protocol}, msg: ${uint8ArrayToString(msg)}`)
         }
       }
     )
@@ -46,19 +43,19 @@ const createNode = async () => {
 
   const { stream: stream1 } = await node1.dialProtocol(node2.peerId, ['/a'])
   await pipe(
-    ['protocol (a)'],
+    [uint8ArrayFromString('protocol (a)')],
     stream1
   )
 
   const { stream: stream2 } = await node1.dialProtocol(node2.peerId, ['/b'])
   await pipe(
-    ['protocol (b)'],
+    [uint8ArrayFromString('protocol (b)')],
     stream2
   )
 
   const { stream: stream3 } = await node1.dialProtocol(node2.peerId, ['/b'])
   await pipe(
-    ['another stream on protocol (b)'],
+    [uint8ArrayFromString('another stream on protocol (b)')],
     stream3
   )
 })();
