@@ -1,25 +1,23 @@
 /* eslint-disable no-console */
-'use strict'
 
-const Libp2p = require('../../')
-const TCP = require('libp2p-tcp')
-const Mplex = require('libp2p-mplex')
-const { NOISE } = require('@chainsafe/libp2p-noise')
-const Gossipsub = require('@achingbrain/libp2p-gossipsub')
-const { fromString: uint8ArrayFromString } = require('uint8arrays/from-string')
-const { toString: uint8ArrayToString } = require('uint8arrays/to-string')
+import { createLibp2p } from 'libp2p'
+import { TCP } from '@libp2p/tcp'
+import { Mplex } from '@libp2p/mplex'
+import { Noise } from '@chainsafe/libp2p-noise'
+import { Gossipsub } from '@achingbrain/libp2p-gossipsub'
+import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
+import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
+import { CustomEvent } from '@libp2p/interfaces'
 
 const createNode = async () => {
-  const node = await Libp2p.create({
+  const node = await createLibp2p({
     addresses: {
       listen: ['/ip4/0.0.0.0/tcp/0']
     },
-    modules: {
-      transport: [TCP],
-      streamMuxer: [Mplex],
-      connEncryption: [NOISE],
-      pubsub: Gossipsub
-    }
+    transports: [new TCP()],
+    streamMuxers: [new Mplex()],
+    connectionEncryption: [new Noise()],
+    pubsub: new Gossipsub()
   })
 
   await node.start()
@@ -35,22 +33,20 @@ const createNode = async () => {
   ])
 
   // Add node's 2 data to the PeerStore
-  await node1.peerStore.addressBook.set(node2.peerId, node2.multiaddrs)
+  await node1.peerStore.addressBook.set(node2.peerId, node2.getMultiaddrs())
   await node1.dial(node2.peerId)
 
-  node1.pubsub.on(topic, (msg) => {
-    console.log(`node1 received: ${uint8ArrayToString(msg.data)}`)
+  node1.pubsub.addEventListener(topic, (evt) => {
+    console.log(`node1 received: ${uint8ArrayToString(evt.detail.data)}`)
   })
-  node1.pubsub.subscribe(topic)
 
   // Will not receive own published messages by default
-  node2.pubsub.on(topic, (msg) => {
-    console.log(`node2 received: ${uint8ArrayToString(msg.data)}`)
+  node2.pubsub.addEventListener(topic, (evt) => {
+    console.log(`node2 received: ${uint8ArrayToString(evt.detail.data)}`)
   })
-  node2.pubsub.subscribe(topic)
 
   // node2 publishes "news" every second
   setInterval(() => {
-    node2.pubsub.publish(topic, uint8ArrayFromString('Bird bird bird, bird is the word!'))
+    node2.pubsub.dispatchEvent(new CustomEvent(topic, { detail: uint8ArrayFromString('Bird bird bird, bird is the word!') }))
   }, 1000)
 })()
