@@ -1,4 +1,3 @@
-'use strict'
 
 process.on('unhandedRejection', (err) => {
   console.error(err)
@@ -6,11 +5,14 @@ process.on('unhandedRejection', (err) => {
   process.exit(1)
 })
 
-const path = require('path')
-const fs = require('fs')
-const {
+import path from 'path'
+import fs from 'fs'
+import {
   waitForOutput
-} = require('./utils')
+} from './utils.js'
+import { fileURLToPath } from 'url'
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
 async function testAll () {
   for (const dir of fs.readdirSync(__dirname)) {
@@ -24,7 +26,7 @@ async function testAll () {
       continue
     }
 
-    await waitForOutput('npm info ok', 'npm', ['test', '--', dir], {
+    await waitForOutput('npm info ok', 'npm', ['--loglevel', 'info', 'run', 'test', '--', dir], {
       cwd: __dirname
     })
   }
