@@ -1,22 +1,23 @@
-'use strict'
+import { TCP } from '@libp2p/tcp'
+import { WebSockets } from '@libp2p/websockets'
+import { Mplex } from '@libp2p/mplex'
+import { Noise } from '@chainsafe/libp2p-noise'
+import defaultsDeep from '@nodeutils/defaults-deep'
+import { createLibp2p as create } from 'libp2p'
 
-const TCP = require('libp2p-tcp')
-const WS = require('libp2p-websockets')
-const mplex = require('libp2p-mplex')
-const { NOISE } = require('@chainsafe/libp2p-noise')
-const defaultsDeep = require('@nodeutils/defaults-deep')
-const libp2p = require('../../..')
-
-async function createLibp2p(_options) {
+export async function createLibp2p(_options) {
   const defaults = {
-    modules: {
-      transport: [TCP, WS],
-      streamMuxer: [mplex],
-      connEncryption: [NOISE],
-    },
+    transports: [
+      new TCP(),
+      new WebSockets()
+    ],
+    streamMuxers: [
+      new Mplex()
+    ],
+    connectionEncryption: [
+      new Noise()
+    ]
   }
 
-  return libp2p.create(defaultsDeep(_options, defaults))
+  return create(defaultsDeep(_options, defaults))
 }
-
-module.exports = createLibp2p
