@@ -1,11 +1,7 @@
-'use strict'
+import { test as test1 } from './test-1.js'
+import { test as test2 } from './test-2.js'
 
-const test1 = require('./test-1')
-const test2 = require('./test-2')
-
-async function test() {
+export async function test() {
   await test1()
   await test2()
 }
-
-module.exports = test
