@@ -1,13 +1,12 @@
-'use strict'
+import path from 'path'
+import { waitForOutput } from '../utils.js'
+import { fileURLToPath } from 'url'
 
-const path = require('path')
-const { waitForOutput } = require('../utils')
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
-async function test () {
+export async function test () {
   await waitForOutput('This message is sent on a private network', 'node', [path.join(__dirname, 'index.js')], {
     cwd: __dirname
   })
 }
-
-module.exports = test
 
