@@ -1,14 +1,13 @@
-'use strict'
+import path from 'path'
+import { waitForOutput } from '../utils.js'
+import { fileURLToPath } from 'url'
 
-const path = require('path')
-const { waitForOutput } = require('../utils')
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
-async function test () {
+export async function test () {
   process.stdout.write('2.js\n')
 
   await waitForOutput('another stream on protocol (b)', 'node', [path.join(__dirname, '2.js')], {
     cwd: __dirname
   })
 }
-
-module.exports = test
