@@ -1,11 +1,12 @@
-'use strict'
+import path from 'path'
+import execa from 'execa'
+import pDefer from 'p-defer'
+import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
+import { fileURLToPath } from 'url'
 
-const path = require('path')
-const execa = require('execa')
-const pDefer = require('p-defer')
-const { toString: uint8ArrayToString } = require('uint8arrays/to-string')
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
-async function test () {
+export async function test () {
   const defer = pDefer()
   process.stdout.write('1.js\n')
 
@@ -26,5 +27,3 @@ async function test () {
   await defer.promise
   proc.kill()
 }
-
-module.exports = test
